Be able to skip chmod'ing /downloads folder
I'd like to be able to skip chmod'ing the /downloads folder in the Docker container. As my downloads folder is actually on my NAS and the `/downloads` folder is exposed through NFS with limited permissions, chmod'ing is not allowed. But I also know that it's not necessary, as I've set `$PUID` and `$GUID` to the correct values.