CMakeLists.txt - Fix omission in CMAKE_CXX_FLAGS_DEBUG (#338)
When building with `-DCMAKE_BUILD_TYPE=Debug`, using clang, the wrong
`CMAKE_C_FLAGS_DEBUG` are set, which causes a wrong debug build. This
issue is caused by 0043452, which sets `CMAKE_CXX_FLAGS_DEBUG` to the
correct flags, but later is set to `CMAKE_CXX_FLAGS`, which is empty.

Issue does not occur when building with a different compiler.