Build: exclude regex when already provided by the system (#589)
nzbget includes a quite old version of the [glibc regex](https://sourceware.org/git/?p=glibc.git;a=blob;f=posix/regex.c;hb=HEAD).
Building and using such old version of regex is unnecessary in most platforms, as the latest and better versions of regex are included in all POSIX distributions for decades.

The old embedded regex is showing various concerning issues:
- potential memory leaks in `re_dfa_add_node` from the `re_realloc` calls
- dereference of null pointer in `create_initial_state`
- dereference of null pointer in `calc_first`
- dereference of null pointer in `duplicate_tree`
- potential memory leaks in `re_copy_regs` from the `re_realloc` calls
- dereference of null pointer in `re_search_internal`
- attempt to free released memory in `prune_impossible_nodes`
- assigned value is garbage or undefined in `build_trtable`
- use of function `alloca` is discouraged; there is no way to check for failure but failure may still occur, resulting in a possibly exploitable security vulnerability

Hence the best initial action is to avoid using this old code altogether when it's not necessary.